Make the content folder structure follow the layout name.
