+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Components/FixedInfo/OrderInfo.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
+using osu.Game.Graphics.Sprites;
+using osu.Game.Rulesets.Karaoke.Objects;
+
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.FixedInfo;
+
+public partial class OrderInfo : OsuSpriteText
+{
+    private readonly IBindable<int> bindableOrder;
+
+    public OrderInfo(Lyric lyric)
+    {
+        bindableOrder = lyric.OrderBindable.GetBoundCopy();
+    }
+
+    [BackgroundDependencyLoader]
+    private void load()
+    {
+        bindableOrder.BindValueChanged(value =>
+        {
+            int order = value.NewValue;
+            Text = $"#{order}";
+        }, true);
+    }
+}
