+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Compose/DetailLyricList.cs
@@ -1,0 +1,54 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Compose;
+
+public partial class DetailLyricList : BaseLyricList
+{
+    public DetailLyricList()
+    {
+        AdjustSkin(skin =>
+        {
+            skin.FontSize = 15;
+        });
+    }
+
+    protected override DrawableLyricList CreateDrawableLyricList()
+        => new DrawableDetailLyricList();
+
+    protected override Drawable CreateBackground(LyricEditorColourProvider colourProvider, LyricEditorMode mode)
+    {
+        bool containsHandler = mode == LyricEditorMode.EditText;
+
+        const float timing_base_width = LYRIC_LIST_PADDING + DetailRow.TIMING_WIDTH;
+        float timingWidth = containsHandler ? DrawableLyricListItem.HANDLER_WIDTH + timing_base_width : timing_base_width;
+        return new GridContainer
+        {
+            ColumnDimensions = new[]
+            {
+                new Dimension(GridSizeMode.Absolute, timingWidth),
+                new Dimension(),
+            },
+            Content = new[]
+            {
+                new[]
+                {
+                    new Box
+                    {
+                        RelativeSizeAxes = Axes.Both,
+                        Colour = colourProvider.Background3(mode),
+                    },
+                    new Box
+                    {
+                        RelativeSizeAxes = Axes.Both,
+                        Colour = colourProvider.Background4(mode),
+                    },
+                },
+            },
+        };
+    }
+}
